general bug fix in distributed algo - sim crashes after 1 minute - code to be cleaned

+++ controllers/supervisor/include/message.h
@@ -31,7 +31,7 @@ typedef struct {
   int event_index;
   // ]]
   int16_t sender_id;
-  float market_bids[NUM_TASKS];
+  int market_bids[NUM_TASKS];
   int16_t market_winners[NUM_TASKS];
 } message_t;
